place interface inside actions file
https://coveord.atlassian.net/browse/KIT-661
