+++ packages/headless/src/features/facets/range-facets/date-facet-set/date-facet-actions.ts
@@ -1,6 +1,5 @@
 import dayjs from 'dayjs';
 import {createAction} from '@reduxjs/toolkit';
-import {DateFacetRegistrationOptions} from './interfaces/options';
 import {DateFacetValue} from './interfaces/response';
 import {updateRangeFacetSortCriterion} from '../generic/range-facet-actions';
 import {
@@ -20,6 +19,71 @@ import {facetIdDefinition} from '../../generic/facet-actions-validation';
 import {RangeFacetSortCriterion} from '../generic/interfaces/request';
 import {dateFacetValueDefinition} from '../generic/range-facet-validate-payload';
 import {buildDateRange} from '../../../../controllers/facets/range-facet/date-facet/date-range';
+import {DateRangeRequest} from './interfaces/request';
+
+export interface DateFacetRegistrationOptions {
+  /**
+   * A unique identifier for the facet.
+   */
+  facetId: string;
+
+  /**
+   * The field whose values you want to display in the facet.
+   */
+  field: string;
+
+  /**
+   * Whether the index should automatically create range values.
+   *
+   * Tip: If you set this parameter to true, you should ensure that the ['Use cache for numeric queries' option](https://docs.coveo.com/en/1982/#use-cache-for-numeric-queries) is enabled for this facet's field in your index in order to speed up automatic range evaluation.
+   */
+  generateAutomaticRanges: boolean;
+
+  /**
+   * The values displayed by the facet in the search interface at the moment of the request.
+   *
+   * If `generateAutomaticRanges` is false, values must be specified.
+   * If `generateAutomaticRanges` is true, automatic ranges are going to be appended after the specified values.
+   *
+   * @defaultValue `[]`
+   */
+  currentValues?: DateRangeRequest[];
+
+  /**
+   * Whether to exclude folded result parents when estimating the result count for each facet value.
+   *
+   * @defaultValue `true`
+   */
+  filterFacetCount?: boolean;
+
+  /**
+   * The maximum number of results to scan in the index to ensure that the facet lists all potential facet values.
+   *
+   * Note: A high injectionDepth may negatively impact the facet request performance.
+   *
+   * Minimum: `0`
+   *
+   * @defaultValue `1000`
+   */
+  injectionDepth?: number;
+
+  /**
+   * The number of values to request for this facet.
+   * Also determines the number of additional values to request each time this facet is expanded, and the number of values to display when this facet is collapsed.
+   *
+   * Minimum: `1`
+   *
+   * @defaultValue `8`
+   */
+  numberOfValues?: number;
+
+  /**
+   * The sort criterion to apply to the returned facet values.
+   *
+   * @defaultValue `ascending`
+   */
+  sortCriteria?: RangeFacetSortCriterion;
+}
 
 const dateRangeRequestDefinition = {
   start: requiredNonEmptyString,
